FESI1-29 [Feat]: ProgressBar 공통 컴포넌트화 & 연산 예외 처리
